refactor: rename prom_read * remove duplicate variables
